+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Drums class to allow easy referencing of MIDI drums by name.
+
 
 
 ## 2.4.0 (2023-06-23)